% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_AF.R
\name{extract_AF}
\alias{extract_AF}
\title{Extract a table for population attributable fraction for logistic regression models}
\usage{
extract_AF(
  object,
  data,
  exposure,
  mark,
  case.control = FALSE,
  confidence.level = 0.95
)
}
\arguments{
\item{object}{A regression object given by glm() function}

\item{data}{An optional data frame, list or environment (or object coercible by as.data.frame
to a data frame) containing the variables in the model. If not found in data, the
variables are taken from environment (formula), typically the environment from
which the function is called.}

\item{exposure}{The name of the exposure variable as a string.}

\item{mark}{A mark created by the user to identify model(s)}

\item{case.control}{Can be set to TRUE if the data is from a non-matched case control study. By
default case.control is set to FALSE which is used for cross-sectional sampling designs.}

\item{confidence.level}{User-specified confidence level for the confidence intervals. If not specified it
defaults to 95 percent. Should be specified in decimals such as 0.95 for 95 percent.}
}
\value{
A table of values for PAF.
}
\description{
This function takes a logistic regression object as an input and returns an
object of PAF values and its 95\% confidence intervals (95\% CIs).
}
\examples{
# Create an example logistic model
example_model <- glm(y ~ x + c1 + c2, data = example_data, family = 'binomial')
return_AF <- extract_AF(object = example_model, data = example_data, exposure = "x", mark = 'example_1')
}
