% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coeff_lm.R
\name{extract_coeff_lm}
\alias{extract_coeff_lm}
\title{Extract coefficients for linear regression models}
\usage{
extract_coeff_lm(regress_obj, index_lst, mark)
}
\arguments{
\item{regress_obj}{An object given by lm() function}

\item{index_lst}{A list of index(s) of variable(s) of interest}

\item{mark}{A mark created by users to identify model(s)}
}
\value{
A table of extracted model values
}
\description{
This function takes a regression object as an input and returns the
extracted values of beta coefficient, 95\% confidence intervals (95\% CIs),
and P value for one or more given variable(s) of each model.
}
\examples{
# Create an example logistic model
example_model <- lm(y ~ x + c1 + c2, data = example_data)

# Extract OR values for x
tab_1 <- extract_coeff_lm(example_model, 2, 'first_example')

# Extract OR values for x and c1
tab_2 <- extract_coeff_lm(example_model, c(2, 3), 'second_example')
}
