% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_OR.R
\name{extract_OR}
\alias{extract_OR}
\title{Extract ORs for logistic regression models}
\usage{
extract_OR(regress_obj, index_lst, mark)
}
\arguments{
\item{regress_obj}{An object given by glm() function}

\item{index_lst}{A list of index(s) of variable(s) of interest}

\item{mark}{A mark created by users to identify model(s)}
}
\value{
A table of extracted model values
}
\description{
This function takes a regression object as an input and returns the
extracted values of beta coefficient, odds ratio (OR), 95\% confidence
intervals (95\% CIs), and P value for one or more given variable(s) of each model.
}
\examples{
# Create an example logistic model
example_model <- glm(y ~ x + c1 + c2, data = example_data, family = 'binomial')

# Extract OR values for x
tab_1 <- extract_OR(example_model, 2, 'first_example')

# Extract OR values for x and c1
tab_2 <- extract_OR(example_model, c(2, 3), 'second_example')
}
