% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mendelian_randomization.R
\name{get_MR_value}
\alias{get_MR_value}
\title{Make a table to extract results from Mendelian randomization}
\usage{
get_MR_value(reg_object1, reg_object2, mark)
}
\arguments{
\item{reg_object1}{A regression model given by glm() function: exposure ~ gene_instrument (+ covariates)}

\item{reg_object2}{A regression model given by glm() function: outcome ~ gene_instrument (+ covariates)}

\item{mark}{Mark of the MR model.}
}
\value{
A table of extracted results from Mendelian randomization
}
\description{
This function takes two glm() models as input: 1) exposure ~ gene_instrument + covariates;
2) outcome ~ gene_instrument + covairates; and returns the extracted values of beta coefficient,
95\% confidence intervals (95\% CIs), and P value for MR analysis.
}
\examples{
# Create two models for MR analysis
f0_exposure <- glm(exposure ~ gene_instrument, data = example_data, family = 'binomial')
f0_outcome <- glm(outcome ~ gene_instrument, data = example_data, family = 'binomial')
crude_mr <- get_MR_value(f0_exposure, f0_outcome, 'example_MR')
}
