% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coeff_lm.R
\name{make_coeff_table_lm}
\alias{make_coeff_table_lm}
\title{Make a coefficient table for multiple linear regression models}
\usage{
make_coeff_table_lm(lst_model, index_lst, model_head)
}
\arguments{
\item{lst_model}{A list of linear regressionmodels}

\item{index_lst}{A list of index(s) of variable(s) of interest}

\item{model_head}{The mark created by users to identify model(s)}
}
\value{
A table of extracted model values for multiple models
}
\description{
This function takes a list of linear regression models as an input and returns the
extracted values of beta coefficient, 95\% confidence intervals (95\% CIs),
and P value for one or more given variable(s) of each model.
}
\examples{
# Create an example linear model lists
lst_model <- c('model_1', 'model_2', 'model_3')

model_1 <- lm(y ~ x, data = example_data)
model_2 <- lm(y ~ x + c1, data = example_data)
model_3 <- lm(y ~ x + c1 + c2, data = example_data)

# Make a coefficient table for variable x
eg_linear_tab <- make_coeff_table_lm(lst_model, 2, 'example_')
}
