% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Make a correlation heat map}
\usage{
make_heatmap(
  dataset,
  sig_level = 0.95,
  lab_size = 3,
  fontsize = 12,
  width = 8,
  height = 8,
  save_name,
  output_path
)
}
\arguments{
\item{dataset}{A dataset of interest (should all be numeric variables).}

\item{sig_level}{Significant level to be marked in the heat map, 0.95 by default.}

\item{lab_size}{Size for label.}

\item{fontsize}{Size for font.}

\item{width}{Output figure width.}

\item{height}{Output figure height.}

\item{save_name}{Saved name for your figure.}

\item{output_path}{File path for output figure.}
}
\value{
A heat map for correlations across all covariates.
}
\description{
This function takes a dataset of interest as an input and returns a heat map showing
Pearson correlation coefficients and their significance. Build based on ggcorrplot and ggplot2.
}
\examples{
make_heatmap(example_corr, sig_level = 0.95, lab_size = 3,
fontsize = 12, width = 9, height = 9, 'example_output.png', output_path)
}
