% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_OR.R
\name{make_OR_table}
\alias{make_OR_table}
\title{Make an OR table for multiple logistic regression models}
\usage{
make_OR_table(lst_model, index_lst, model_head)
}
\arguments{
\item{lst_model}{A list of logistic regressionmodels}

\item{index_lst}{A list of index(s) of variable(s) of interest}

\item{model_head}{The mark created by users to identify model(s)}
}
\value{
A table of extracted model values for multiple models
}
\description{
This function takes a list of logistic regression models as an input and returns the
extracted values of beta coefficient, odds ratio (OR), 95\% confidence
intervals (95\% CIs), and P value for one or more given variable(s) of each model.
}
\examples{
# Create an example logistic model lists
lst_model <- c('model_1', 'model_2', 'model_3')

model_1 <- glm(y ~ x, data = example_data, family = 'binomial')
model_2 <- glm(y ~ x + c1, data = example_data, family = 'binomial')
model_3 <- glm(y ~ x + c1 + c2, data = example_data, family = 'binomial')

# Make an OR table for variable x
eg_OR_tab <- make_OR_table(lst_model, 2, 'example_')
}
