% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mendelian_randomization.R
\name{MR_assumption_test}
\alias{MR_assumption_test}
\title{Test for Mendelian randomization assumptions}
\usage{
MR_assumption_test(
  gene_instrument,
  exposure,
  outcome,
  covariates = NULL,
  data_name
)
}
\arguments{
\item{gene_instrument}{The gene instrument used in the MR analysis (should be single SNP/polygenic score).}

\item{exposure}{Exposure of the MR model.}

\item{outcome}{Outcome of the MR model.}

\item{covariates}{Covariates adjusted in the model (default as NULL).}

\item{data_name}{Name of the dataset. Please notice that the dataset should have NO missing value.}
}
\value{
A set of TRUE/FALSE indicating if the assumptions hold. If the MR assumptions hold,
all value should be TRUE.
}
\description{
This function could be used for testing MR assumptions. Rules suggested by Glymour et al.
See details at https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3366596/pdf/kwr323.pdf.
}
\examples{
MR_assumption_test('gene_instrument', 'exposure', 'outcome', covariates = 'a + b', 'example_dataset')
}
